requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.5
Django==4.0.6
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.7.1
flake8==4.0.1
identify==2.5.2
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.10
virtualenv==20.15.1
